Convert floats to doubles

+++ WormholeRenderer/WormholeRenderer.cpp
@@ -6,15 +6,15 @@
 #include <iostream>
 #include <fstream>
 
-float PI = 3.1415926535897f;
+double PI = 3.1415926535897;
 
 using namespace ATP::Wormhole::Ellis;
 
 int main()
 {
 	ATP::Math::Vector camera(1, 2, 3);	//Camera location
-	float psi, theta;					//Viewing direction
-	float roll;							//Camera roll
+	double psi, theta;					//Viewing direction
+	double roll;							//Camera roll
 
 	//--Begin metric test trace--
 		//Open output file
@@ -24,11 +24,11 @@ int main()
 		//Iterate until heading directly away from the wormhole
 		outFile << "p,t,dp,dt,r,x,y,z\n";
 
-		float p = -10.0f;			//Location of camera in radial coordinates
-		float t = PI;				//Location of camera in radial coordinates
-		float T = 0.2507875f * PI;	//Viewing angle
-		float w = 1.0f;				//Wormhole throat width
-		float ds = 0.01f;			//Integration step size
+		double p = -10.0;			//Location of camera in radial coordinates
+		double t = PI;				//Location of camera in radial coordinates
+		double T = 0.2507875 * PI;	//Viewing angle
+		double w = 1.0;				//Wormhole throat width
+		double ds = 0.01;			//Integration step size
 
 		Geodesic g(p, t, T, w);
